requirements.txt
Flask==3.1.2
flask-cors==6.0.1
google-generativeai==0.8.5
pytesseract==0.3.13
Pillow==11.3.0
pdfplumber==0.11.7
requests==2.32.5
flask-cors
python-dotenv
nltk
scikit-learn
PyMuPDF
